Rename core/common/hash_value.h to core/common/basic_types.h.
